Update README.md
An your linux machine try running cowsay Hi There!, but it is not installed.  Write an ansible playbook to install the 'cowsay' program on the linux host using the apt module.  Run it with ansible-playbook and verify it works by running the 'cowsay' program.  Add it to the crontab running every 5 minutes. Once you see it run (check ansible.log and syslog), delete the 'cowsay' program with sudo apt purge cowsay.  Verify you cannot run it.  Wait until the playbook runs again and they try the cowsay program again.  It should work.

To turn this in copy these items and their output to a text file and submit it via git:

the configmgmt3.yml script
the crontab file showing the script setup to run every 5 minutes
the sectio of the ansible.log file showing it ran and reinstalled cowsay